medbridge/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-medbridge"
version = "0.1.0"
description = "MedBridge - A medical communication RL environment that trains LLMs to explain complex diagnoses to patients in their native Indian language"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Reward function dependencies
    "textstat>=0.7.0",
    "langdetect>=1.0.9",
    "transformers>=4.40.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m medbridge.server.app
server = "medbridge.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["medbridge", "medbridge.server"]
package-dir = { "medbridge" = ".", "medbridge.server" = "server" }